Refactor UI
