+++ apps/frontend/components/judging/judge-advisor/awards-panel.tsx
@@ -8,7 +8,6 @@ import {
   WSClientEmittedEvents,
   WSServerEmittedEvents
 } from '@lems/types';
-import ResultExportPaper from './result-export-paper';
 import DeliberationsPaper from './deliberations-paper';
 import AwardsPaper from './awards-paper';
